Refactor: Separate class for auto indexing JSON
