+++ fooddata_vegattributes/abstract_indexed_fooddata.py
@@ -3,8 +3,8 @@ from typing import Iterable, Union
 
 from .fooddata import FoodDataDict
 from .utils.abstract_indexed_json import AbstractIndexedJson
+from .utils.auto_indexing_json import AutoIndexingJsonableWriter, IndexSpec
 from .utils.close_on_exit import CloseOnExit
-from .utils.indexed_json_utils import IndexingIterator, IndexSpec
 
 
 class AbstractIndexedFoodDataJson(CloseOnExit, metaclass=ABCMeta):
@@ -14,34 +14,26 @@ class AbstractIndexedFoodDataJson(CloseOnExit, metaclass=ABCMeta):
   def close(self): ...
 
   def write_fooddata_dicts(self, ds: Iterable[FoodDataDict]):
-    indexed_ds = IndexingIterator(
-      elements=ds,
-      primary_index=IndexSpec("fdc-id", lambda d: str(d["fdcId"])),
-      secondary_indices=[
-        IndexSpec(
-          "ingredient-code",
-          lambda d: str(d.get("foodCode") or d.get("ndbNumber")),
-        ),
-        IndexSpec(
-          "fdc-category-description",
-          lambda d: (
-            d.get("wweiaFoodCategory", {}).get("wweiaFoodCategoryDescription")
-            or d.get("foodCategory", {}).get("description")
+    auto_indexing_writer: AutoIndexingJsonableWriter[FoodDataDict] = (
+      AutoIndexingJsonableWriter(
+        self.indexed_json,
+        primary_index=IndexSpec("fdc-id", lambda d: str(d["fdcId"])),
+        secondary_indices=[
+          IndexSpec(
+            "ingredient-code",
+            lambda d: str(d.get("foodCode") or d.get("ndbNumber")),
           ),
-        ),
-      ]
-    )
-    self.indexed_json.write_jsonables(
-      indexed_ds.primary_index.name, indexed_ds
-    )
-    for name, map in indexed_ds.secondary_maps.items():
-      self.indexed_json.write_links(
-        name,
-        (
-          (k, [(indexed_ds.primary_index.name, target) for target in targets])
-          for k, targets in map.items()
-        ),
+          IndexSpec(
+            "fdc-category-description",
+            lambda d: (
+              d.get("wweiaFoodCategory", {}).get("wweiaFoodCategoryDescription")
+              or d.get("foodCategory", {}).get("description")
+            ),
+          ),
+        ]
       )
+    )
+    auto_indexing_writer.write_jsonables(ds)
 
   def get_fooddata_dict_by_fdc_id(
     self, fdc_id: Union[int, str]
